PoC for reset and close buttons
