+++ app/styles/less/modules/DataGrid.less
@@ -70,28 +70,6 @@
 .data-grid td.show-more::before { display: none; }
 .data-grid-resizer { position: absolute; top: 0px; bottom: 0px; width: 5px; z-index: 500; }
 
-.data-grid {
-    * { min-width: 0px; min-height: 0px; }
-    * { box-sizing: border-box; }
-    :focus { outline-width: 0px; }
-    input[type="radio"]:focus { outline: -webkit-focus-ring-color auto 5px; }
-    img { -webkit-user-drag: none; }
-    iframe, a img { border: none; }
-    button, input, select { font-family: inherit; font-size: inherit; }
-    input { background-color: white; color: inherit; }
-    input::placeholder { color: rgba(0, 0, 0, 0.54); }
-    span[is="dt-icon-label"] { flex: 0 0 auto; }
-    [is="ui-icon"] { display: inline-block; flex-shrink: 0; }
-    [is="ui-icon"].icon-mask { background-color: rgb(110, 110, 110); -webkit-mask-position: var(--spritesheet-position); }
-    .spritesheet-smallicons:not(.icon-mask) { background-image: url("/images/smallIcons.svg"); }
-    .spritesheet-smallicons.icon-mask { -webkit-mask-image: url("/images/smallIcons.svg"); }
-    .spritesheet-largeicons.icon-mask { -webkit-mask-image: url("Images/largeIcons.svg"); }
-    .spritesheet-mediumicons.icon-mask { -webkit-mask-image: url("Images/mediumIcons.svg"); }
-    :host-context(.force-white-icons) [is="ui-icon"].spritesheet-smallicons, .force-white-icons [is="ui-icon"].spritesheet-smallicons, [is="ui-icon"].force-white-icons.spritesheet-smallicons, -theme-preserve { -webkit-mask-image: url("/images/smallIcons.svg"); -webkit-mask-position: var(--spritesheet-position); background: rgb(250, 250, 250) !important; }
-    ::selection { background-color: rgb(187, 222, 251); }
-    [data-aria-utils-animation-hack] { animation: 0s ease 0s 1 normal none running ANIMATION-HACK; }
-    cursor: default; font-family: ".SFNSDisplay-Regular", "Helvetica Neue", "Lucida Grande", sans-serif; font-size: 12px; tab-size: 4; user-select: none; color: rgb(34, 34, 34); background: white;
-}
 
 :root { height: 100%; overflow: hidden; }
 :root { --accent-color: #1a73e8; --accent-fg-color: #1a73e8; --accent-color-hover: #3b86e8; --active-control-bg-color: #5a5a5a; --focus-bg-color: hsl(214, 40%, 92%); --input-validation-error: #db1600; --toolbar-bg-color: #f3f3f3; --toolbar-hover-bg-color: #eaeaea; --selection-fg-color: white; --selection-inactive-fg-color: #5a5a5a; --selection-inactive-bg-color: #dadada; --tab-selected-fg-color: #333; --tab-selected-bg-color: var(--toolbar-bg-color); --drop-shadow: 0 0 0 1px rgba(0, 0, 0, 0.05), 0 2px 4px rgba(0, 0, 0, 0.2), 0 2px 6px rgba(0, 0, 0, 0.1); --divider-color: #d0d0d0; --focus-ring-inactive-shadow: 0 0 0 1px #e0e0e0; --item-selection-bg-color: #cfe8fc; --item-selection-inactive-bg-color: #e0e0e0; }
