+++ app/styles/less/modules/ODataView.less
@@ -88,7 +88,7 @@ odata-detail-view {
         background-color: Lightblue
     }
 
-    .hidden {
+    div.hidden {
         visibility: hidden;
     }
 
@@ -134,7 +134,7 @@ odata-detail-view {
 
     .data-grid, #editor, .editorAlt {
         position: absolute;
-        top: 1.3rem;
+        top: 25px;
         left: 0;
         right: 0;
         bottom: 0;
@@ -159,5 +159,88 @@ odata-detail-view {
         a {
             color: lightblue;
         }
+    }
+}
+
+
+// from devTools (TODO: refactor and add copyright)
+.odataTab {
+    * { min-width: 0px; min-height: 0px; }
+    * { box-sizing: border-box; }
+    :focus { outline-width: 0px; }
+    input[type="radio"]:focus { outline: -webkit-focus-ring-color auto 5px; }
+    img { -webkit-user-drag: none; }
+    iframe, a img { border: none; }
+    button, input, select { font-family: inherit; font-size: inherit; }
+    input { background-color: white; color: inherit; }
+    input::placeholder { color: rgba(0, 0, 0, 0.54); }
+    span[is="dt-icon-label"] { flex: 0 0 auto; }
+    [is="ui-icon"] { display: inline-block; flex-shrink: 0; }
+    [is="ui-icon"].icon-mask { background-color: rgb(110, 110, 110); -webkit-mask-position: var(--spritesheet-position); }
+    .spritesheet-smallicons:not(.icon-mask) { background-image: url("/images/smallIcons.svg"); }
+    .spritesheet-smallicons.icon-mask { -webkit-mask-image: url("/images/smallIcons.svg"); }
+    .spritesheet-largeicons.icon-mask { -webkit-mask-image: url("/images/largeIcons.svg"); }
+    .spritesheet-mediumicons.icon-mask { -webkit-mask-image: url("/images/mediumIcons.svg"); }
+    .spritesheet-mediumicons:not(.icon-mask) { background-image: url(/images/mediumIcons.svg); }
+    :host-context(.force-white-icons) [is="ui-icon"].spritesheet-smallicons, .force-white-icons [is="ui-icon"].spritesheet-smallicons, [is="ui-icon"].force-white-icons.spritesheet-smallicons, -theme-preserve { -webkit-mask-image: url("/images/smallIcons.svg"); -webkit-mask-position: var(--spritesheet-position); background: rgb(250, 250, 250) !important; }
+    ::selection { background-color: rgb(187, 222, 251); }
+    [data-aria-utils-animation-hack] { animation: 0s ease 0s 1 normal none running ANIMATION-HACK; }
+    cursor: default; font-family: ".SFNSDisplay-Regular", "Helvetica Neue", "Lucida Grande", sans-serif; font-size: 12px; tab-size: 4; user-select: none; color: rgb(34, 34, 34); background: white;
+
+    .largeicon-clear:hover:not(:active) { background-color: #333; }
+
+    [is=ui-icon]:not(.icon-mask) {
+        background-position: var(--spritesheet-position);
+    }
+
+    .close-button {
+        width: 14px;
+        height: 14px;
+        cursor: default;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
+
+    .hover-icon, .active-icon {
+        display: none;
+    }
+
+    .close-button:hover .default-icon, .close-button:active .default-icon {
+        display: none;
+    }
+
+    .close-button:hover .hover-icon {
+        display: block;
+    }
+
+    .close-button[data-keyboard-focus="true"]:focus .default-icon, .close-button:active .default-icon {
+        display: none;
+    }
+
+    .close-button[data-keyboard-focus="true"]:focus .hover-icon {
+        display: block;
+    }
+
+    .close-button:active .hover-icon {
+        display: none !important;
+    }
+
+    .close-button:active .active-icon {
+        display: block;
+    }
+
+    .toolbar-item {
+        position: relative;
+        display: flex;
+        background-color: transparent;
+        flex: none;
+        align-items: center;
+        justify-content: center;
+        padding: 0;
+        height: 26px;
+        width: 26px;
+        border: none;
+        white-space: pre;
     }
 }
